+++ tests/resources/testdb.sql
@@ -1000,7 +1000,6 @@ CREATE TABLE `template_wells` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 
-
 --
 -- Table structure for table `transfer_plates`
 --