+++ tests/test_connection.py
@@ -1,10 +1,14 @@
+import json
 import os
+from pathlib import Path
 
 import pytest
+from peewee import InterfaceError
 
 from valarpy import *
 
 CONFIG_PATH = Path(__file__).parent / "resources" / "connection.json"
+CONFIG_DATA = json.loads(CONFIG_PATH.read_text(encoding="utf8"))
 
 
 class TestModel:
@@ -16,12 +20,20 @@ class TestModel:
             valar.reconnect()
             assert list(Refs) is not None
 
-    def test_config_path(self):
+    def test_config_dict(self):
+        with Valar(CONFIG_DATA):
+            from valarpy.model import Refs
+
+            list(Refs.select())
+
+    def test_config_path_env(self):
         popped = None
         try:
+            # just in case it's needed elsewhere, let's save the current VALARPY_CONFIG
             if "VALARPY_CONFIG" in os.environ:
                 popped = os.environ.pop("VALARPY_CONFIG")
-            with pytest.raises(LookupError):
+            # test using environment variable
+            with pytest.raises(FileNotFoundError):
                 with Valar():
                     from valarpy.model import Refs
 
@@ -34,6 +46,16 @@ class TestModel:
                     list(Refs.select())
             os.environ["VALARPY_CONFIG"] = str(CONFIG_PATH)
             with Valar():
+                from valarpy.model import Refs
+
+                assert list(Refs) is not None
+            # test with passing directly
+            os.environ.pop("VALARPY_CONFIG")
+            with Valar(CONFIG_PATH):
+                from valarpy.model import Refs
+
+                assert list(Refs) is not None
+            with Valar([Path("asd346erfdgawq046j54e4y"), CONFIG_PATH]):
                 from valarpy.model import Refs
 
                 assert list(Refs) is not None