+++ CHANGELOG.md
@@ -3,10 +3,23 @@
 Adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [2.0.x] - 2020-08-14
+
+## [2.1.0] - 2020-08-25
 
 ### Fixed:
-- Slightly incorrect docs reqs
+- Incompatibility with real database
+- Improved connection code
+- Inflexible dependency version ranges
+- Poor code separation: moved general code from `model.py` to `metamodel.py`
+
+### Removed:
+- `DagsToCreate` and `GeneticConstructs`, which were invalid
+
+### Added:
+- `new_model` and `opened` context managers
+- Test that checks each model class
+- Support for `.chemfish/connection.json`
+- Better main method, useful for testing with the active db
 
 
 ## [2.0.0] - 2020-08-14